Initial checkin of KV78turbo OVapi
This exists out of three components

The UWSGI client which receives HTTP requests via UWSGI, transmits the request via zeroMQ to the API which replies back with the result
The main API which receives KV8 data from OpenOV. The received messages are parsed from CTX and put in the datastore dictionaries. When there is no extra information source from Koppelvlak7 stored, it will query the Postgresql server for that static information. When the KV8 passtime is not in the KV7 database it will print an error, this signals either an outdated KV7 database or missing KV7 packages, in the latter case contact OpenOV. In most cases there will be only a few queries like that depending on the amount of KV7 passes you cache. As of what's in the repo only stops which departe 60 minutes later than now will be queried.
The KV7 component, which each minute queries the database for the selected timerange in the future and transmits this data via zeroMQ push/pull to the main API.

Manual:

uwsgi --http://127.0.0.1:10078 -w kv78turbo-client
python kv78turbo-api.py
python kv7-db.py

Requirements:

A postgresql filled up with KV7 data in KV7 database scheme. Make sure to create the indices with indices.sql to obtain the necessary performance